chore: Extract index tests
