+++ packages/cli/src/__tests__/ceramic-daemon.test.ts
@@ -12,8 +12,6 @@ import {
   TimedAbortSignal,
   GenesisCommit,
   TestUtils,
-  CommitType,
-  StreamState,
 } from '@ceramicnetwork/common'
 import { TileDocument } from '@ceramicnetwork/stream-tile'
 import { firstValueFrom } from 'rxjs'
@@ -27,12 +25,11 @@ import { makeCeramicCore } from './make-ceramic-core.js'
 import { makeCeramicDaemon } from './make-ceramic-daemon.js'
 
 const seed = 'SEED'
-const MODEL_STREAM_ID = new StreamID(1, TestUtils.randomCID())
 
 describe('Ceramic interop: core <> http-client', () => {
   jest.setTimeout(30000)
   let ipfs: IpfsApi
-  let tmpFolder: any
+  let tmpFolder: tmp.DirectoryResult
   let core: Ceramic
   let daemon: CeramicDaemon
   let client: CeramicClient
@@ -47,7 +44,7 @@ describe('Ceramic interop: core <> http-client', () => {
 
   beforeEach(async () => {
     tmpFolder = await tmp.dir({ unsafeCleanup: true })
-    core = await makeCeramicCore(ipfs, tmpFolder.path, [MODEL_STREAM_ID].map(String))
+    core = await makeCeramicCore(ipfs, tmpFolder.path)
     daemon = await makeCeramicDaemon(core)
     const apiUrl = `http://localhost:${daemon.port}`
     client = new CeramicClient(apiUrl, { syncInterval: 500 })
@@ -628,89 +625,6 @@ describe('Ceramic interop: core <> http-client', () => {
       // Now force re-pin and make sure underlying state and ipfs records get re-pinned
       await client.pin.add(docA.id, true)
       expect(pinSpy).toBeCalledTimes(4)
-    })
-  })
-
-  describe('index api', () => {
-    describe('getCollection calls IndexAPI', () => {
-      test('model in query', async () => {
-        const query = new URL(`http://localhost:${daemon.port}/api/v0/collection`)
-        query.searchParams.set('model', MODEL_STREAM_ID.toString())
-        query.searchParams.set('first', '100')
-        const indexSpy = jest.spyOn(daemon.ceramic.index, 'queryIndex')
-        await fetchJson(query.toString())
-        expect(indexSpy).toBeCalledWith({
-          first: 100,
-          model: MODEL_STREAM_ID,
-        })
-      })
-      test('too much entries requested: forward pagination', async () => {
-        const query = new URL(`http://localhost:${daemon.port}/api/v0/collection`)
-        query.searchParams.set('model', MODEL_STREAM_ID.toString())
-        query.searchParams.set('first', '20000')
-        await expect(fetchJson(query.toString())).rejects.toThrow(
-          /Requested too many entries: 20000/
-        )
-      })
-      test('too much entries requested: forward pagination', async () => {
-        const query = new URL(`http://localhost:${daemon.port}/api/v0/collection`)
-        query.searchParams.set('model', MODEL_STREAM_ID.toString())
-        query.searchParams.set('last', '20000')
-        await expect(fetchJson(query.toString())).rejects.toThrow(
-          /Requested too many entries: 20000/
-        )
-      })
-      test('model, account in query', async () => {
-        const query = new URL(`http://localhost:${daemon.port}/api/v0/collection`)
-        query.searchParams.set('model', MODEL_STREAM_ID.toString())
-        const account = `did:key:${randomString(10)}`
-        query.searchParams.set('account', account)
-        query.searchParams.set('first', '100')
-        const indexSpy = jest.spyOn(daemon.ceramic.index, 'queryIndex')
-        await fetchJson(query.toString())
-        expect(indexSpy).toBeCalledWith({
-          first: 100,
-          model: MODEL_STREAM_ID,
-          account: account,
-        })
-      })
-      test('serialize StreamState', async () => {
-        const query = new URL(`http://localhost:${daemon.port}/api/v0/collection`)
-        query.searchParams.set('model', MODEL_STREAM_ID.toString())
-        query.searchParams.set('first', '100')
-        const original = daemon.ceramic.index.queryIndex.bind(daemon.ceramic.index)
-        const fauxStreamState = {
-          type: 0,
-          log: [
-            {
-              type: CommitType.GENESIS,
-              cid: TestUtils.randomCID(),
-            },
-          ],
-        } as unknown as StreamState
-        // Return faux but serializable StreamState
-        daemon.ceramic.index.queryIndex = async () => {
-          return {
-            edges: [
-              {
-                cursor: 'opaque-cursor',
-                node: fauxStreamState,
-              },
-            ],
-            pageInfo: {
-              hasNextPage: false,
-              hasPreviousPage: false,
-            },
-          }
-        }
-        // It gets serialized
-        const response = await fetchJson(query.toString())
-        expect(response.edges.length).toEqual(1)
-        // Check if it is indeed the same serialized state
-        expect(response.edges[0].node).toEqual(StreamUtils.serializeState(fauxStreamState))
-        // Get the original queryIndex method back
-        daemon.ceramic.index.queryIndex = original
-      })
     })
   })
 })
